minor config change (l-out)

+++ admin/expert/index.php
@@ -41,7 +41,7 @@ require_once('../config/version.php');
                 $is_paused = glob('/etc/*_paused');
                 $repl_str = array('/\/etc\//', '/_paused/');
                 $paused_modes = preg_replace($repl_str, '', $is_paused);
-                if (!empty($is_paused)) {
+                if (!empty($is_paused) && $_GET['force'] != true) {
                     //HTML output starts here
                     include './header-menu-disabled.inc';
                     echo '<div class="contentwide">'."\n";
